tests were changed and sent to another folder
